scripts/requirements.txt
kafka-python
requests
datasets
pyarrow
